Update flow-binding.yaml
We should use a build-time property in order the Quarkus build to take it into account. I've tested and identified the class that should not be removed, though we can leave it with the wider scope (commented) in case we want to demonstrate a more general way of solving it.